Update and rename ftp-covert-channel.py to ftp-covert.py
Put in some work on the program. Currently, it can log into jeangourd.com, get the listings, and log back out. This is then written to a .txt file, which is then read out to the command line before the program completes. Now, I know it still needs work, but I'm turning in for tonight. I'll pick it up again tomorrow. Until then, feel free to test/add to it.