Update and rename ftp-covert.py to ftp.py
Added more functionality. The only things left to do are to make sure I'm getting the right sequence of 1's and 0's; to make sure the program can decode binary; and to print the decoded message to stdout. If anyone else would like to test/work on it, feel free. After 3:30 today, I won't be able to work on this until I get back on Sunday night.