Update ftp.py
this can be our competition program. I now have to go and get rid of ALL the file open/close stuff so that Gourd will accept it. I went and talked to him, and he doesn't want anything that messes with files, so now I get to go make more string lists.