Create ftp-for-hw
So, the other file I had on here won't fly for Gourd because of file functions I had on there. This is a new and improved version where I replaced the file things with lists. It all works; feel free to download and test. I'd like to upload it during or after class tomorrow so that we have FTP done and over with.