Add files via upload
Modified ftp program where the login, password, and directory name are provided on the command line in addition to python ftp-cyber.py. The only thing that ever has to be changed is the hardcoded default to return the result of using all of the permission for binary. It helps to know what directory you want to get into first :)