Chat Timing
This is just more of a starter, with a list of objectives to complete and some components that we'll need to use. If previous programs are used as reference, then most of this program should come together easily. For now, just start on getting the basics to work until Gourd's site comes back up.