xor.py
This is what I have so far for xor.py. It's a work in progress, but I can say this much: The Princess Bride is an excellent movie.

Also, the keys and cipheretext have a lot of space in them. I need to account for that, but so far it can do the xor part.